Refactor avatar component into shared
