+++ web/src/app/shared/components/header-layout/current-user-widget/current-user-widget.module.ts
@@ -16,7 +16,7 @@
 
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { UserAvatarModule } from 'app/components/user-avatar/user-avatar.module';
+import { UserAvatarModule } from 'app/shared/components/user-avatar/user-avatar.module';
 import { MatButtonModule } from '@angular/material/button';
 import { CurrentUserWidgetComponent } from 'app/shared/components/header-layout/current-user-widget/current-user-widget.component';
