+++ web/src/app/shared/components/header-layout/current-user-widget/user-profile-popup/user-profile-popup.module.ts
@@ -19,7 +19,7 @@ import { CommonModule } from '@angular/common';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialogModule } from '@angular/material/dialog';
 import { UserProfilePopupComponent } from 'app/shared/components/header-layout/current-user-widget/user-profile-popup/user-profile-popup.component';
-import { UserAvatarModule } from 'app/components/user-avatar/user-avatar.module';
+import { UserAvatarModule } from 'app/shared/components/user-avatar/user-avatar.module';
 
 @NgModule({
   declarations: [UserProfilePopupComponent],
