+++ web/src/app/shared/components/user-avatar/user-avatar.component.html
