+++ web/src/app/shared/components/user-avatar/user-avatar.component.spec.ts
@@ -16,7 +16,7 @@
 
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
-import { UserAvatarComponent } from 'app/components/user-avatar/user-avatar.component';
+import { UserAvatarComponent } from 'app/shared/components/user-avatar/user-avatar.component';
 
 describe('UserAvatarComponent', () => {
   let component: UserAvatarComponent;
